Create interface to extract class contract
Following [SOLID](https://en.wikipedia.org/wiki/SOLID) principles, start
separating contract from implementation. In order to have more modular
code it is good to go "interface oriented", so we can expose the "what"
without leaking the "how".

First of all, an interface file is added. The name of the component is
assigned to the interface, which will represent the functionality. It's
concrete implementation is renamed.