Avoid leaking implementation details
The fact that the ParserFacade class uses a File internally to offer IO
capabilities must be encapsulated. Users of objects of this class don't
need to know that detail. So the design to create an instance of those
objects is modified in the following way:

* No getter / setter methods for the internal File object. The file will
be pass on the instance at construction time, with no chance to modify
it or call operations on it without ParserFacade knowledge. If a new
Facade for a new File is required, a new Facade instance will be used.
Each instance will handle only one file. That makes our Facade object in
some way immutable (cannot operate on a different file from the one
assigned when instantiated). Immutability of objects is a best practice
preached since the beginning of time (even by Joshua Bloch, contributor
to the standard Java library, back in 2001). In order to enforce that
immutability inside the class as well, both the constructor parameter
and class attribute are declared final.

* A package private constructor, that prevents the direct instantiation
of the ParserFacadeImpl class from outside it's package (putting it
inside a named package will be done in a future commit).

* The static factory method is now implemented on the interface that
will expose the contract for the Facade as a default method. As the
interface lives in the same package that it's implementation it has the
required visibility to proxy it's instantiation. The return type of the
new factory method is ParserFacade (the interface). That isolates the
implementation from the rest of the code that will use it.

* Our ParserFacade is not a singleton anymore, and should not be. If we
want it to encapsulate it's internal state (a File object reference) a
different instance for each file will be required.