Move code base to a named package
It is a bad practice to use the default package. Might be Ok for a code
test like this one; but due to the design applied in the previous
commits (that relies on package access modifiers for a class and it's
constructor) we create a named package for this code. That way we will
be able to make more obvious how the pubic interface of this module will
become.