Remove code duplication in get operations
A common pattern can be observed in both getXXX methods
(ParserFacadeImpl class). It can be abstracted out using a high order
function. Thanks to the fancy Java 8 _java.util.function_ package we
have the Predicate interface at our disposal. We use it to encode the
condition that varies inside the get methods.