Optimize String concatenation using StringBuilder
In Java the String class is immutable. That means that every time we
concatenate two Strings three instances are created, and two of them
discarded.

Using a mutable utility such as StringBuilder makes our code less
expensive in terms of object allocation and garbage collection. As long
as that mutability does not leak outside the method that uses it, it's
ok to use mutable data structures to improve performance without
compromising code quality.