Add test for ParserFacade
Usually I'm used to apply TDD, so instead of writing the test as the
last step I usually write it upfront. Here we already had code to
refactor, so I've implemented the test after the refactor is completed.

The best way to do this is to add the test before refactoring, so we can
be sure that our refactor didn't break anything.